Added displayImageName to JSON
The block library will now show a display image from the json, rather than showing exactly what appears on screen. This allows alternative/cleaner pictures to be used for the block library.

If a display image is not specified, the standard on-screen image will be used. Therefore, a display image is optional